LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-transition-mismatched-list-expected.txt


FAIL A custom property of type <angle># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <angle>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <color># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <color>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <custom-ident># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <custom-ident>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <image># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <image>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <integer># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <integer>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <length-percentage># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <length-percentage>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <length># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <length>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <number># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <number>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <percentage># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <percentage>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <resolution># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <resolution>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <time># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <time>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <url># yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
FAIL A custom property of type <url>+ yields a discrete CSS Transition if the lists do not contain the same number of values assert_equals: A single animation is running expected 1 but got 0
